+++ apps/opentelemetry_api/lib/open_telemetry.ex
@@ -117,8 +117,6 @@ defmodule OpenTelemetry do
 
   defdelegate get_tracer(name), to: :opentelemetry
   defdelegate get_tracer(name, vsn, schema_url), to: :opentelemetry
-  defdelegate register_tracer(name, vsn), to: :opentelemetry
-  defdelegate register_tracer(name, vsn, schema_url), to: :opentelemetry
   defdelegate set_default_tracer(t), to: :opentelemetry
 
   # Helpers to build OpenTelemetry structured types